update certificate function

+++ server/truffle/contracts/Transaction.sol
@@ -2,17 +2,17 @@
 pragma solidity >=0.4.22 <0.9.0;
 
 contract Transaction {
-    mapping(uint => string) transactions;
+    mapping(uint256 => string) transactions;
 
-    function addTransaction(uint txId, string memory _hash) public {
+    function addTransaction(uint256 txId, string memory _hash) public {
         transactions[txId] = _hash;
     }
 
-    function getHashByTransactionId(uint txId)
+    function getHashByTransactionId(uint256 txId)
         public
         view
         returns (string memory)
     {
-      return transactions[txId];
+        return transactions[txId];
     }
 }
